backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
torch>=2.0.0
torchvision>=0.15.0
timm>=1.0.0
opencv-python-headless==4.8.1.78
albumentations>=1.4.0
Pillow>=10.0.0
numpy>=1.26.4

